MOVED StudyPlan into ./study_plan
* Since StudyPlan is not a common nor top-level component, put it in a
specific folder
* Ensured dependencies file paths in other affect components that mount
StudyPlan reflect the change in file paths
